add initial plans

+++ internal/core/plan.go
@@ -9,7 +9,7 @@ import (
 type Plan struct {
 	ID            uuid.UUID `db:"id"`
 	Name          string    `db:"name"`
-	Price         int       `db:"price"`
+	Price         float64   `db:"price"`
 	StripePriceID string    `db:"stripe_price_id"`
 	CreatedAt     time.Time `db:"created_at"`
 	UpdatedAt     time.Time `db:"updated_at"`
